Fix data loss in connection with geo: intents.
When geo: intents are handled by activities created
in a different task, that can lead to the same map
being opened multiple times (once in each task).
The way the maps are implemented, this means that the
last to call onPause/onDestroy will "win" its reference
points will end up on disk.
This is particularly annoying if that one is the one
that has no reference points and you lost all
that you added...
Solve it via a horrible hack that always uses a new, clean
task (of which there should be only one).